修复AWVS报错

+++ code/app/model/AwvsModel.php
@@ -12,25 +12,28 @@ class AwvsModel extends BaseModel
         $awvs_url = ConfigModel::value('awvs_url');
         $awvs_token = ConfigModel::value('awvs_token');
 
+        if (empty($awvs_url) || empty($awvs_token)) {
+            addlog(["执行AWVS扫描任务失败,未找到有效得配置信息", $awvs_url, $awvs_token]);
+            return false;
+        }
         while (true) {
-            $list = Db::table('app')->whereTime('awvs_scan_time', '<=', date('Y-m-d H:i:s', time() - (86400 * 15)))->where('is_delete',0)->limit(1)->orderRand()->select()->toArray();
+            $list = Db::table('app')->whereTime('awvs_scan_time', '<=', date('Y-m-d H:i:s', time() - (86400 * 15)))->where('is_delete', 0)->limit(1)->orderRand()->select()->toArray();
             foreach ($list as $val) {
                 $url = $val['url'];
                 $id = $val['id'];
                 if (filter_var($url, FILTER_VALIDATE_URL) === false) {
                     addlog(["URL地址不正确", $id, $url]);
-                    self::scanTime('app',$id,'awvs_scan_time');
+                    self::scanTime('app', $id, 'awvs_scan_time');
                     continue;
                 }
-                addlog(["AWVS开始执行扫描任务", $id, $url]);
                 //添加目标
-                $targetId = self::getTargetId($id, $url,$awvs_url,$awvs_token,$val['user_id']);
+                $targetId = self::getTargetId($id, $url, $awvs_url, $awvs_token, $val['user_id']);
                 if (!$targetId) {
                     addlog(["AWVS扫描失败", $id, $url]);
-                    self::scanTime('app',$id,'awvs_scan_time');
+                    self::scanTime('app', $id, 'awvs_scan_time');
                     continue;
                 }
-                $retArr = self::getScanStatus($targetId,$awvs_url,$awvs_token);
+                $retArr = self::getScanStatus($targetId, $awvs_url, $awvs_token);
                 if (isset($retArr['code']) && $retArr['code'] == 404) {
                     addlog(["未在AWVS中找到此目标ID", $targetId]);
                     Db::table('awvs_app')->where(['target_id' => $targetId])->delete();
@@ -42,20 +45,20 @@ class AwvsModel extends BaseModel
                 }
                 //判断目标扫描状态
                 if (isset($retArr['last_scan_session_status']) && $retArr['last_scan_session_status'] == 'completed') {
-                    self::addVulnList($retArr['last_scan_id'], $retArr['last_scan_session_id'],$awvs_url,$awvs_token,$val['user_id']);
-                    self::scanTime('app',$id,'awvs_scan_time');
+                    self::addVulnList($retArr['last_scan_id'], $retArr['last_scan_session_id'], $awvs_url, $awvs_token, $val['user_id']);
+                    self::scanTime('app', $id, 'awvs_scan_time');
                 }
             }
-            addlog("AWS累了，休息10秒钟...");
-            sleep(10);
+            addlog("AWVS累了，休息30秒钟...");
+            sleep(30);
         }
     }
 
-    public static function addVulnList($scanId, $scanSessionId,$awvs_url,$awvs_token,$user_id)
+    public static function addVulnList($scanId, $scanSessionId, $awvs_url, $awvs_token, $user_id)
     {
-        $vulnList = self::getVulnList($scanId, $scanSessionId,$awvs_url,$awvs_token);
+        $vulnList = self::getVulnList($scanId, $scanSessionId, $awvs_url, $awvs_token);
         foreach ($vulnList['vulnerabilities'] as $value) {
-            $detail = self::getDetail($scanId, $scanSessionId, $value['vuln_id'],$awvs_url,$awvs_token);
+            $detail = self::getDetail($scanId, $scanSessionId, $value['vuln_id'], $awvs_url, $awvs_token);
             $value = array_merge($value, $detail);
             foreach ($value as $k => $v) {
                 $value[$k] = is_string($v) ? $v : json_encode($v, JSON_UNESCAPED_UNICODE);
@@ -65,7 +68,7 @@ class AwvsModel extends BaseModel
         }
     }
 
-    public static function getDetail($scanId, $scanSessionId, $vulnId,$awvs_url,$awvs_token)
+    public static function getDetail($scanId, $scanSessionId, $vulnId, $awvs_url, $awvs_token)
     {
         $ch = curl_init();
 
@@ -89,7 +92,7 @@ class AwvsModel extends BaseModel
         return json_decode($result, true);
     }
 
-    public static function getVulnList($scanId, $scanSessionId,$awvs_url,$awvs_token)
+    public static function getVulnList($scanId, $scanSessionId, $awvs_url, $awvs_token)
     {
         $ch = curl_init();
 
@@ -114,7 +117,7 @@ class AwvsModel extends BaseModel
 
     }
 
-    public static function startScan($targetId,$awvs_url,$awvs_token)
+    public static function startScan($targetId, $awvs_url, $awvs_token)
     {
         $postData = "{\"profile_id\":\"11111111-1111-1111-1111-111111111119\",\"schedule\":{\"disable\":false,\"start_date\":null,\"time_sensitive\":false},\"target_id\":\"{$targetId}\"}";
         $ch = curl_init();
@@ -140,7 +143,7 @@ class AwvsModel extends BaseModel
 
     }
 
-    public static function getScanStatus($targetId,$awvs_url,$awvs_token)
+    public static function getScanStatus($targetId, $awvs_url, $awvs_token)
     {
         $ch = curl_init();
 
@@ -165,7 +168,7 @@ class AwvsModel extends BaseModel
         return $retArr;
     }
 
-    public static function getTargetId($id, $url,$awvs_url,$awvs_token,$user_id)
+    public static function getTargetId($id, $url, $awvs_url, $awvs_token, $user_id)
     {
         $appInfo = Db::table('awvs_app')->where(['app_id' => $id])->find();
         if (empty($appInfo)) {
@@ -183,7 +186,7 @@ class AwvsModel extends BaseModel
             $result = curl_exec($ch);
             if (curl_errno($ch)) {
                 echo 'Error:' . curl_error($ch);
-                return  false;
+                return false;
             }
             curl_close($ch);
             $appInfo = json_decode($result, true);
@@ -193,7 +196,7 @@ class AwvsModel extends BaseModel
                 Db::table('awvs_app')->insert($appInfo);
             }
             //添加扫描任务
-            self::startScan($appInfo['target_id']);
+            self::startScan($appInfo['target_id'], $awvs_url, $awvs_token);
         }
         return $appInfo['target_id'];
     }
